Created new script that manages the health
I have created a new script called GlobalHealthManager which is going to be used by every object that has health (buildings, ai, player). Also modified the script of each of the mentioned entities to adapt to this new script. This change was made in order to have an easier time on GUI part of things, the objective health part.